+++ packages/assets/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @webpack-blocks/assets - Changelog
 
+## 1.0.0-rc
+
+- Added a `styleLoader` option to `css()` and `css.modules()` blocks.
+With it it's now possible to pass options to the `style-loader` or
+remove it from loaders completely.
+
 ## 1.0.0-beta.2
 
 - Update dependency versions