+++ packages/extract-text/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @webpack-blocks/extract-text - Changelog
 
+## 1.0.0-rc
+
+- Breaking change: drop webpack 2 support, update extract-text-webpack-plugin to v3
+
 ## 1.0.0-beta.2
 
 - Update dependency version